backend/requirements.txt
Flask
python-dotenv
flask-session
authlib
pywebpush
flask-sqlalchemy
psycopg2-binary